requirements.txt
# LLM API integration (required for automated benchmarking via OpenRouter)
openai>=1.0.0
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Web API server
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.11.0

# For formatting and colored output (optional)
# rich>=13.0.0
